Refactor bot into modular application package
